Extract Task into its own module
