+++ lib/run/frame.ts
@@ -1,21 +1,20 @@
 import type {
   Frame,
-  Future,
   Instruction,
   Operation,
-  Reject,
   Resolve,
   Result,
-  Task,
 } from "../types.ts";
 
-import { type Computation, evaluate, reset, shift } from "../deps.ts";
+import { evaluate, reset, shift } from "../deps.ts";
 import { shiftSync } from "../shift-sync.ts";
 import { lazy } from "../lazy.ts";
-
-import { createValue } from "./value.ts";
-import { create } from "./create.ts";
 import { Err, Ok } from "../result.ts";
+
+import type { Exit, FrameResult } from "./types.ts";
+import { createValue } from "./value.ts";
+import { createTask } from "./task.ts";
+import { create } from "./create.ts";
 
 let ids = 0;
 
@@ -187,16 +186,6 @@ export function createFrame<T>(options: FrameOptions<T>): Frame<T> {
   return frame!;
 }
 
-type Exit<T> = {
-  type: "aborted";
-} | {
-  type: "crashed";
-  error: Error;
-} | {
-  type: "result";
-  result: Result<T>;
-};
-
 type Thunk = ReturnType<typeof $next>;
 
 // deno-lint-ignore no-explicit-any
@@ -231,110 +220,3 @@ type InstructionResult =
   | {
     type: "interrupted";
   };
-
-type FrameResult<T> = Result<void> & {
-  exit: Exit<T>;
-};
-
-import { action } from "../instructions.ts";
-
-function createTask<T>(
-  frame: Frame<T>,
-  results: Computation<FrameResult<T>>,
-): Task<T> {
-  let promise: Promise<T>;
-
-  let awaitResult = (resolve: Resolve<T>, reject: Reject) => {
-    evaluate(function* () {
-      let result = getResult(yield* results);
-
-      if (result.ok) {
-        resolve(result.value);
-      } else {
-        reject(result.error);
-      }
-    });
-  };
-
-  let getPromise = () => {
-    promise = new Promise<T>((resolve, reject) => {
-      awaitResult(resolve, reject);
-    });
-    getPromise = () => promise;
-    return promise;
-  };
-
-  let task = create<Task<T>>("Task", {}, {
-    *[Symbol.iterator]() {
-      let frameResult = evaluate<FrameResult<T> | void>(() => results);
-      if (frameResult) {
-        let result = getResult(frameResult);
-        if (result.ok) {
-          return result.value;
-        } else {
-          throw result.error;
-        }
-      } else {
-        return yield* action<T>(function* (resolve, reject) {
-          awaitResult(resolve, reject);
-        });
-      }
-    },
-    then: (...args) => getPromise().then(...args),
-    catch: (...args) => getPromise().catch(...args),
-    finally: (...args) => getPromise().finally(...args),
-    halt() {
-      let haltPromise: Promise<void>;
-      let getHaltPromise = () => {
-        haltPromise = new Promise((resolve, reject) => {
-          awaitHaltResult(resolve, reject);
-        });
-        getHaltPromise = () => haltPromise;
-        frame.destroy();
-        return haltPromise;
-      };
-      let awaitHaltResult = (resolve: Resolve<void>, reject: Reject) => {
-        evaluate(function* () {
-          let result = yield* results;
-          if (result.ok) {
-            resolve(result.value);
-          } else {
-            reject(result.error);
-          }
-        });
-      };
-      return create<Future<void>>("Future", {}, {
-        *[Symbol.iterator]() {
-          let result = evaluate<FrameResult<T> | void>(() => results);
-
-          if (result) {
-            if (!result.ok) {
-              throw result.error;
-            }
-          } else {
-            yield* action<void>(function* (resolve, reject) {
-              awaitHaltResult(resolve, reject);
-              frame.destroy();
-            });
-          }
-        },
-        then: (...args) => getHaltPromise().then(...args),
-        catch: (...args) => getHaltPromise().catch(...args),
-        finally: (...args) => getHaltPromise().finally(...args),
-      });
-    },
-  });
-  return task;
-}
-
-function getResult<T>(result: FrameResult<T>): Result<T> {
-  if (!result.ok) {
-    return result;
-  } else if (result.exit.type === "aborted") {
-    return Err(Error("halted"));
-  } else if (result.exit.type === "crashed") {
-    return Err(result.exit.error);
-  } else {
-    return result.exit.result;
-  }
-}
